various(fix): fixed various problems in the app
