+++ src/config/index.js
@@ -2,5 +2,6 @@ export const CONFIG = {
   appName: 'Investment Programming & Management System',
   shortName: 'IPMS',
   author: 'Mark Lester A. Bolotaolo',
-  graphQLEndpoint: 'http://localhost:8000/graphql'
+  // graphQLEndpoint: 'http://localhost:8000/graphql'
+  graphQLEndpoint: 'https://wh490179.ispot.cc/graphql'
 }
