+++ src/modules/profile/store/actions.js
@@ -1,18 +1,18 @@
 import { profileService } from '@/services/profile.service'
 
-export function profile() {
-	profileService.getCurrentUser()
+export function profile({}) {
+	return profileService.getCurrentUser()
 }
 
-export function updateProfile(payload) {
-	profileService.updateProfile(payload)
+export function updateProfile({}, payload) {
+	return profileService.updateProfile(payload)
 }
 
-export function uploadUserAvatar(payload) {
-	profileService.uploadUserAvatar(payload)
+export function uploadUserAvatar({}, payload) {
+	return profileService.uploadUserAvatar(payload)
 }
 
 export function chooseAvatar({}, payload) {
 	console.log(payload)
-	profileService.chooseAvatar(payload)
+	return profileService.chooseAvatar(payload)
 }
