+++ src/modules/projects/pages/ProjectsPage.vue
@@ -4,30 +4,39 @@
       <q-btn color="primary" label="search" @click="searchProjectDialog = true"></q-btn>
     </page-title>
 
-    <template v-if="paginatedProjects.data && paginatedProjects.data.length">
-
-      <project-pagination 
-        v-model="currentPage" 
-        :max="paginatedProjects.paginatorInfo ? paginatedProjects.paginatorInfo.lastPage: 1" 
-        :max-pages="10" 
-        @input="setPage" />
-
-      <template v-for="project in paginatedProjects.data">
-        <project-item :project="project" :key="project.id"></project-item>
-      </template>
-
-      <project-pagination 
-        v-model="currentPage" 
-        :max="paginatedProjects.paginatorInfo ? paginatedProjects.paginatorInfo.lastPage: 1" 
-        :max-pages="10" 
-        @input="setPage" />
-
+    <template v-if="$apollo.loading">
+      <div>
+        Loading...
+      </div>
     </template>
 
     <template v-else>
-      <div class="flex flex-center text-primary">
-        Loading content...
-      </div>
+      <template v-if="paginatedProjects.data">
+
+        <template v-if="paginatedProjects.data.length">
+          <project-pagination 
+            v-model="currentPage" 
+            :max="paginatedProjects.paginatorInfo ? paginatedProjects.paginatorInfo.lastPage: 1" 
+            :max-pages="10" 
+            @input="setPage" />
+
+          <template v-for="project in paginatedProjects.data">
+            <project-item :project="project" :key="project.id"></project-item>
+          </template>
+
+          <project-pagination 
+            v-model="currentPage" 
+            :max="paginatedProjects.paginatorInfo ? paginatedProjects.paginatorInfo.lastPage: 1" 
+            :max-pages="10" 
+            @input="setPage" />
+        </template>
+
+        <div v-else>
+          <no-item icon="cancel" message="No project found."></no-item>
+        </div>
+
+      </template>
+
     </template>
 
     <search-project 
@@ -58,6 +67,7 @@ import PageContainer from '@/ui/page/PageContainer'
 import ProjectPagination from '../components/ProjectPagination'
 import SearchProject from '../components/dialogs/SearchProject'
 import { Loading } from 'quasar'
+import NoItem from '@/ui/components/NoItem'
 
 const PER_PAGE = 10
 
@@ -68,8 +78,20 @@ export default {
 		PageTitle,
 		ProjectItem,
     ProjectPagination,
-    SearchProject
+    SearchProject,
+    NoItem
 	},
+  apollo: {
+    paginatedProjects: {
+      query: PAGINATED_PROJECTS,
+      variables() {
+        return {
+          first: PER_PAGE,
+          page: this.currentPage
+        }
+      }
+    }
+  },
   data() {
     return {
       paginatedProjects: {},
@@ -82,10 +104,11 @@ export default {
   },
   watch: {
     $route(to, from) {
+      
       console.log('i am triggered from route change')
       // set currentPage to the query page of destination
       // this will update the query since the variables is reactive
-      this.loadProjects(parseInt(to.query.page))
+      // this.loadProjects(parseInt(to.query.page))
     }
   },
   computed: {
@@ -93,22 +116,11 @@ export default {
   },
   methods: {
     setPage(pageNumber) {
-      console.log(`setPage: ${pageNumber}`)
 
       // this will change the route based on the incoming pageNumber
-      this.$router.push({ name: 'index-project', query: { page: pageNumber } }).catch(() => { console.log('changed route from setPage')})
-    },
-    loadProjects(pageNumber) {
-      Loading.show()
-      this.$store.dispatch('projects/fetchProjects', {
-        first: 10,
-        page: pageNumber
-      })
-      .then(res => {
-        this.paginatedProjects = res.paginatedProjects
-
-        Loading.hide()
-      })
+      this
+        .$router
+        .push({ name: 'index-project', query: { page: pageNumber } })
     }
   },
   created() {
@@ -118,7 +130,7 @@ export default {
 
     this.currentPage = currentPage;
 
-    this.loadProjects(currentPage)
+    // this.loadProjects(currentPage)
   }
 };
 </script>
